Overall Changes
Added Music and new items, added an end game feature with a game timer, updated player movement so the player is reset to a standing position after walking. Added simple UI. I will update game rendering soon and add NPCs i just didn't get to that today. Most of this is transferable to the project we need to turn in I Just need to figure out how.